+++ bonsoir/lib/src/broadcast.dart
@@ -1,43 +1,23 @@
+import 'package:bonsoir/src/action_handler.dart';
 import 'package:bonsoir_platform_interface/bonsoir_platform_interface.dart';
 import 'package:flutter/foundation.dart';
 
 /// Allows to broadcast a service on the network.
-class BonsoirBroadcast {
+class BonsoirBroadcast extends BonsoirActionHandler<BonsoirBroadcastEvent> {
   /// The service to broadcast.
   final BonsoirService service;
-
-  /// The event source abstraction
-  final BonsoirAction<BonsoirBroadcastEvent> _broadcastAction;
 
   /// Creates a new Bonsoir broadcast instance.
   BonsoirBroadcast({
     bool printLogs = kDebugMode,
     required this.service,
-  }) : _broadcastAction = BonsoirPlatformInterface.instance
-            .createBroadcastAction(service, printLogs: printLogs);
-
-  /// The ready getter, that returns when the platform is ready for broadcast.
-  Future<void> get ready async => _broadcastAction.ready;
-
-  /// This returns whether the platform is ready for broadcast.
-  bool get isReady => _broadcastAction.isReady;
-
-  /// Returns whether the broadcast has been stopped.
-  bool get isStopped => _broadcastAction.isStopped;
-
-  /// Starts the broadcast.
-  Future<void> start() => _broadcastAction.start();
-
-  /// Stops the broadcast.
-  Future<void> stop() => _broadcastAction.stop();
-
-  /// Regular event stream.
-  Stream<BonsoirBroadcastEvent>? get eventStream =>
-      _broadcastAction.eventStream;
+  }) : super(
+          action: BonsoirPlatformInterface.instance.createBroadcastAction(service, printLogs: printLogs),
+        );
 
   @protected
   Map<String, dynamic> toJson() => {
-        ..._broadcastAction.toJson(),
+        ...super.toJson(),
         ...service.toJson(),
       };
 }