+++ bonsoir/lib/src/discovery.dart
@@ -1,40 +1,44 @@
+import 'package:bonsoir/src/action_handler.dart';
 import 'package:bonsoir_platform_interface/bonsoir_platform_interface.dart';
 import 'package:flutter/foundation.dart';
 
 /// Allows to run a network discovery.
-class BonsoirDiscovery {
+class BonsoirDiscovery extends BonsoirActionHandler<BonsoirDiscoveryEvent> {
   /// The type of service to find.
   final String type;
 
-  /// The event source abstraction.
-  final BonsoirAction<BonsoirDiscoveryEvent> _discoveryAction;
+  /// The current service resolver.
+  final ServiceResolver serviceResolver;
+
+  /// Creates a new Bonsoir discovery instance with the given action.
+  BonsoirDiscovery._internal({
+    required this.type,
+    required this.serviceResolver,
+    required super.action,
+  });
 
   /// Creates a new Bonsoir discovery instance.
-  BonsoirDiscovery({
+  factory BonsoirDiscovery({
     bool printLogs = kDebugMode,
-    required this.type,
-  }) : _discoveryAction = BonsoirPlatformInterface.instance
-            .createDiscoveryAction(type, printLogs: printLogs);
-
-  /// The ready getter, that returns when the platform is ready for discovery.
-  Future<void> get ready async => _discoveryAction.ready;
-
-  /// This returns whether the platform is ready for discovery.
-  bool get isReady => _discoveryAction.isReady;
-
-  /// Returns whether the discovery has been stopped.
-  bool get isStopped => _discoveryAction.isStopped;
-
-  /// Starts the discovery.
-  Future<void> start() => _discoveryAction.start();
-
-  /// Stops the discovery.
-  Future<void> stop() => _discoveryAction.stop();
-
-  /// Regular event stream.
-  Stream<BonsoirDiscoveryEvent>? get eventStream =>
-      _discoveryAction.eventStream;
+    required String type,
+    ServiceResolver? serviceResolver,
+  }) {
+    BonsoirAction<BonsoirDiscoveryEvent> action = BonsoirPlatformInterface.instance.createDiscoveryAction(type, printLogs: printLogs);
+    return BonsoirDiscovery._internal(
+      type: type,
+      serviceResolver: action is ServiceResolver ? (action as ServiceResolver) : _NoServiceResolver(),
+      action: action,
+    );
+  }
 
   /// This returns a JSON representation of this discovery.
-  Map<String, dynamic> toJson() => _discoveryAction.toJson()..['type'] = type;
+  Map<String, dynamic> toJson() => super.toJson()..['type'] = type;
+}
+
+/// Created and used as a service resolver if none is provided.
+class _NoServiceResolver with ServiceResolver {
+  @override
+  Future<void> resolveService(BonsoirService service) {
+    throw UnimplementedError('No service resolver provided. Please ensure either you or your current platform interface provide one !');
+  }
 }