+++ bonsoir_platform_interface/lib/src/actions/action.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 import 'dart:math';
 
 import 'package:bonsoir_platform_interface/src/events/event.dart';
+import 'package:bonsoir_platform_interface/src/service/service.dart';
 import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
@@ -36,7 +37,7 @@ abstract class MethodChannelBonsoirAction<T extends BonsoirEvent> extends Bonsoi
   static const String _channelName = 'fr.skyost.bonsoir';
 
   /// The channel.
-  static const MethodChannel channel = MethodChannel(_channelName);
+  static const MethodChannel _channel = MethodChannel(_channelName);
 
   /// The class identifier.
   final int _id;
@@ -68,7 +69,7 @@ abstract class MethodChannelBonsoirAction<T extends BonsoirEvent> extends Bonsoi
   /// Await this method to know when the plugin will be ready.
   @override
   Future<void> get ready async {
-    await channel.invokeMethod('$_classType.initialize', toJson());
+    await _channel.invokeMethod('$_classType.initialize', toJson());
     _eventStream = EventChannel('$_channelName.$_classType.$_id').receiveBroadcastStream().map(transformPlatformEvent);
   }
 
@@ -86,13 +87,13 @@ abstract class MethodChannelBonsoirAction<T extends BonsoirEvent> extends Bonsoi
     assert(isReady, '''$runtimeType should be ready to start in order to call this method.
 You must wait until this instance is ready by calling "await $runtimeType.ready".
 If you have previously called "$runtimeType.stop()" on this instance, you have to create a new instance of this class.''');
-    return channel.invokeMethod('$_classType.start', toJson());
+    return _channel.invokeMethod('$_classType.start', toJson());
   }
 
   /// Stops the current discover or broadcast.
   @override
   Future<void> stop() async {
-    await channel.invokeMethod('$_classType.stop', toJson());
+    await _channel.invokeMethod('$_classType.stop', toJson());
     _isStopped = true;
   }
 
@@ -107,6 +108,16 @@ If you have previously called "$runtimeType.stop()" on this instance, you have t
         'id': _id,
         'printLogs': printLogs,
       };
+
+  /// Invokes a method on the method channel.
+  @protected
+  Future<R?> invokeMethod<R>(String method, [Map<String, dynamic>? arguments]) => _channel.invokeMethod<R>(
+        '$_classType.stop',
+        {
+          ...toJson(),
+          if (arguments != null) ...arguments,
+        },
+      );
 
   /// Allows to generate a random identifier.
   static int _createRandomId() => Random().nextInt(100000);
@@ -147,4 +158,10 @@ mixin AutoStopBonsoirAction<T extends BonsoirEvent> on BonsoirAction<T> {
       stop();
     }
   }
+}
+
+/// An action that is capable of resolving a service.
+mixin ServiceResolver {
+  /// Allows to resolve a service.
+  Future<void> resolveService(BonsoirService service);
 }